[Feat] #5 CardFront와 CardBack 각각의 뷰로 나눔
